+++ back/app/controllers/web_api/v1/nav_bar_items_controller.rb
@@ -12,11 +12,12 @@ class WebApi::V1::NavBarItemsController < ApplicationController
   def removed_default_items
     authorize NavBarItem
     used_codes = NavBarItem.distinct.pluck(:code)
+    rejected_codes = (used_codes + NavBarItemPolicy.feature_disabled_codes).uniq
     @items = NavBarItemService.new.default_items.reject do |item|
       # Not using set difference to have an
       # explicit guarantee of preserving the
       # ordering.
-      used_codes.include? item.code
+      rejected_codes.include? item.code
     end
     render json: WebApi::V1::NavBarItemSerializer.new(@items, params: fastjson_params).serialized_json
   end